Update styling and add new theme colors
The changes in this commit include updates to the styles in Tooltip, Article, Header, and post/page components and the addition of new colors to the tailwind.config.js file. The style updates were made to improve visibility and readability. New theme colors 'dark' for background and 'border' were added to provide more design consistencies. These will be used in the future.