Add new functionality to display Article component correctly.
Refreshed layout of the Article component display to include metadata and descriptions. Three new articles were also added to the content posts.

Fixed an inconsistency in the Article component's layout that caused the metadata and descriptions to be displayed incorrectly. I adjusted the padding and spacing to align the text correctly and allow all information to be visible and clear. These changes are aimed to improve the readability and overall user experience of the site. In addition, three new articles were added to the list of blog posts, which are displayed in the posts directory.