"Refactor navigation items into separate components"
Created separate components - MobileNavItems and HeaderNavItems - to better organize and reuse the navigation items. The elements from the MobileHeaderMenu and Header were extracted into these new components respectively.
Also, the search button in HeaderIcons was updated from an anchor to a button, with an added onClick event, preparing the app for a future implementation of a search feature.