Add custom styles to prose for better readability
Two additional styles are applied to the 'prose' className in 'page.tsx' for better readability. The changes now make sure that the border and background color adapt to the dark mode, while additionally changing the color of the text inside preformatted text - '<pre>' tags. The changes overall improve the visual appearance and usability of the website in both light and dark mode.