Add reading time to article pages
An enhancement to the article pages was made by adding a 'reading time' feature. The function 'calculateReadingTime' was created in 'utils/article.ts' to calculate and return the estimated reading time based on an average reading pace of 200 words per minute. This function was then imported into 'page.tsx' and '[slug]/page.tsx', and used to display the reading time on each individual article page. In addition, the date of publication was also added, using the format function from the date-fns library.