Add Metadata and improve accessibility
Added Metadata to the Page.tsx file and added a method to generate metadata dynamically. This will improve the SEO of the website by providing relevant data to search engines. Also, added accessible labels to the icons in the Footer.tsx file. This will help visually impaired users better understand the page layout.