Add sitemap.xml and update robots.txt for better SEO
New sitemap.xml file was added to improve the SEO of the site, by making it easier for search engines to find and index the pages of the website. The robots.txt file was also updated to include the link to the new sitemap. This will aid in better visibility and search-ability of the site by search engines.