Add Algolia Autocomplete Version 1.2.2 to yarn.lock.
Implemented Algolia Autocomplete Version 1.2.2 into a project. This version provides numerous features and improvements, including faster search and improved accessibility. These changes are expected to provide users with a smoother browsing experience and easier navigation. This upgrade is necessary to keep up to date with the latest features and improvements and enhance the overall user experience. Made changes to the src/App files and CSS styling to accommodate the new feature.