Update explanation of hoisting in JavaScript blog post
Enhanced the clarity of the description of variable hoisting behavior in JavaScript. This change was made in order to provide a more precise articulation on what occurs when a variable is accessed before it's defined. The modification highlights the fact that a 'ReferenceError' is thrown due to absence of the variable in the memory component of the execution context, rather than simply being 'undefined'.